Initial code push
A mixture of code that i have created in different formats which needs to be tested on a physical system.

Currently, xPlanarNormalSequence program works and can be run inside of MAIN without issue. It will go through a sequence of movements designed for an xPlanar system with 2 tiles and 2 movers.